Simplify ec_point_to_octets()
This had an extra dance to allow a NULL output buffer. The plan was to
use this in i2o_ECPublicKey() to preserve the behavior of avoiding an
allocation if out == NULL. However, when I rewrote the latter I punted
on preserving that complication, as it was already batshit crazy enough.

Thus, remove said dance and make ec_point_to_octets() cleaner.

ok jsing