New manual page EVP_aes_128_gcm(3).
The main benefit is moving the cumbersome and error-prone method of
using EVP_EncryptInit(3) for AES-GCM out of the important, but obese
manual page EVP_EncryptInit(3), and to create a logical place for
pointing readers to the safer and more flexible EVP_AEAD_CTX_init(3).

As a side benefit, document three control commands that were so far
undocumented and make the description of three others more precise.

Feedback and OK tb@.