sys/uvideo: forward error bit to consumer
Some invalid frames or frames with an error flag from the camera might be
essential for a stream consumer for synchronization and other purposes.
Without this, the consumer may be unable to synchronize the stream or play
it at all.

Instead of skipping frames that the driver considers incorrect, this change
forwards them to the V4L consumer with the V4L2_BUF_FLAG_ERROR flag set,
allowing the consumer to decide whether to skip the frame or use parts of it.

The behavior for non-mmap consumers remains unchanged, and error frames
are still skipped by the driver.

This change fixes the integrated camera on:
 - ThinkPad T14 Gen 5
 - ThinkPad X1 nano 2
 - Lenovo x13

Plus made progress with support of Elgato HD60 S+

OK: mglocker@