Remove socket state and options checks from the unsplicing path of the sosplice().
The unsplicing paht was the part of the splicing so it followed these
checks too. Socket state and options checks were copy-pasted during
sosplice() reordering to avoid possible API break. However, the splicing
state was never checked, so the unsplicing of non-spliced socket was
always successful. Regarding on this, these checks are useless, moreover
the removal doesn't break the kern/sosplice regression test, so this API
change should be transparent.

Real reason is the simplification of socket usplicing which relies on
socket buffers locks, so there is no reason to lock the socket and stop
packets processing just to do useless checks.

ok bluhm