Use acpipci(4) on hypervisors
If the hypervisor cpuid bit is set, use acpipci to attach PCI busses.
As virtualization is not that old, we can assume that in VMs we don't
need the quirk for old, broken ACPI.

This solves problems with PCI BAR access and recent seabios versions
on qemu.

Agreement from @kettenis