vmm(4) provides VMSA for AMD SEV-ES
For SEV-ES the hypervisor has to provide a VM Save Area (VMSA).  It
will be used by the CPU to load and store encrypted guest CPU state.
Also supply a helper function to retrieve the physical address of
the VMSA.  Non SEV-ES VMs are not affected by this change.

from hshoexer@; OK dv@